RCMS
Robot Controll Mangment system
is used for contolling the robots every move. It recives data over seial and then does the heavy work on computer before sending it back to the arduino as command for the motors. Will also have the ablitly to communicate with websockets for nice display of variables.